refactor: update file name of error.ts to errors.ts for consistency
