+++ packages/taquito-local-forging/test/taquito-local-forging.spec.ts
@@ -7,7 +7,7 @@ import {
   Uint8ArrayConsumer,
 } from '../src/taquito-local-forging';
 import { commonCases } from '../../../integration-tests/data/allTestsCases';
-import { InvalidOperationSchemaError, UnsupportedOperationError } from '../src/error';
+import { InvalidOperationSchemaError, UnsupportedOperationError } from '../src/errors';
 import { InvalidBlockHashError, InvalidOperationKindError } from '@taquito/core';
 import { schemaDecoder, SeedNonceRevelationSchema } from '../src/schema/operation';
 import { ProtoInferiorTo } from '../src/protocols';
